# docs/openapi.yaml
openapi: 3.0.3
info:
    title: Goveto Edge Control API
    version: 0.1.0
    description: Control-plane API for managing edge clusters, nodes, sites, certificates, publish, purge and analytics.
servers:
    - url: /
...
paths:
...
    /api/v1/clusters/{cluster_id}/analytics/summary:
        get:
            operationId: summary
            summary: Analytics summary
            description: Aggregate request/traffic summary for a time range (from/to).
            tags:
                - analytics
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: from
                  in: query
                  schema:
                    type: string
                    format: date-time
                    example: "2026-01-02T15:04:05Z"
                - name: site_id
                  in: query
                  schema:
                    type: string
                - name: to
                  in: query
                  schema:
                    type: string
                    format: date-time
                    example: "2026-01-02T15:04:05Z"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
...
    /api/v1/clusters/{cluster_id}/dns/jobs:
        get:
            operationId: listJobs
            summary: get /api/v1/clusters/{cluster_id}/dns/jobs
            tags:
                - clusters-dns
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
...
    /api/v1/clusters/{cluster_id}/publish/events:
        get:
            operationId: clusterEvents
            summary: Cluster publish events
            description: SSE stream of cluster publish sync status snapshots.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
    /api/v1/clusters/{cluster_id}/publish/status:
        get:
            operationId: clusterStatus
            summary: Cluster publish status
            description: Aggregate publish job counts and recent tasks for the cluster.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        additionalProperties:
                                            type: object
...
    /api/v1/clusters/{cluster_id}/sites/{site_id}/publish:
        get:
            operationId: listJobs
            summary: List publish jobs
            description: List recent publish jobs for a site.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        post:
            operationId: enqueue
            summary: Enqueue publish
            description: Enqueue a site configuration publish job.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "202":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
    /api/v1/clusters/{cluster_id}/sites/{site_id}/publish/{job_id}:
        get:
            operationId: getJob
            summary: Get publish job
            description: Get a single site publish job by id.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: job_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        additionalProperties:
                                            type: object
    /api/v1/clusters/{cluster_id}/sites/{site_id}/purge:
        get:
            operationId: list
            summary: List purge jobs
            description: List recent cache purge jobs for a site.
            tags:
                - purge
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        post:
            operationId: enqueue
            summary: Enqueue purge
            description: Enqueue a cache purge job for the site.
            tags:
                - purge
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                type:
                                    $ref: '#/components/schemas/model_PurgeType'
                                value:
                                    type: string
            responses:
                "202":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        additionalProperties:
                                            type: object
...
components:
    schemas:
...
        model_PurgeType:
            type: string
...
    securitySchemes:
        cookie_session:
            type: apiKey
            in: cookie
            name: session